# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Pikshipp API
  description: Multi-courier shipping aggregator for India
  version: "0.1.0"
  contact:
    name: Pikshipp Engineering
    email: engineering@pikshipp.com
servers:
  - url: https://api.pikshipp.com/v1
    description: Production
  - url: http://localhost:8081/v1
    description: Development
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: pikshipp_session
...
paths:
...
  /webhooks/carriers/{carrier}:
    post:
      summary: Carrier tracking webhook receiver
      tags: [Webhooks]
      security: []
      parameters:
        - name: carrier
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Acknowledged }
        "401": { description: Invalid signature }
...